move CIC specialization to problem file
