+++ src/problems/Gravity3D/gravity_3d.cpp
@@ -63,6 +63,86 @@ template <> struct Physics_Traits<BinaryOrbit> {
 	static constexpr double radiation_constant = 1.0;
 };
 
+namespace quokka {
+// Specialization for CIC particles
+template <> struct ParticleCreationTraits<ParticleType::CIC> {
+	// Specialized nested ParticleChecker for CIC particles
+	template <typename problem_t> struct ParticleChecker {
+		amrex::Real param1 = particle_param1;
+		amrex::Real param2 = particle_param2;
+
+		AMREX_GPU_DEVICE auto operator()(amrex::Array4<const amrex::Real> const &state_arr, int i, int j, int k,
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx, amrex::Real current_time, amrex::Real dt) const -> bool
+		{
+			// A simple demonstration of particle creation
+			// Could check density threshold or other state-based conditions
+			amrex::ignore_unused(state_arr, dx);
+			const int spacing = 16;
+			const bool is_create_particle_1 = current_time <= param1 && current_time + dt > param1;
+			const bool is_create_particle_2 = current_time <= param2 && current_time + dt > param2;
+			return (is_create_particle_1 || is_create_particle_2) && (i != 0 && i % spacing == 0) && (j != 0 && j % spacing == 0) &&
+			       (k != 0 && k % spacing == 0);
+		}
+	};
+
+	// Specialized nested ParticleCreator for CIC particles
+	template <typename problem_t> struct ParticleCreator {
+		int mass_idx;
+		int cpu_id;
+		amrex::Long pid_start;
+		amrex::Real param1 = particle_param1;
+		amrex::Real param2 = particle_param2;
+
+		AMREX_GPU_HOST_DEVICE
+		ParticleCreator(int mass_index, int processor_id, amrex::Long particle_id_start)
+		    : mass_idx(mass_index), cpu_id(processor_id), pid_start(particle_id_start)
+		{
+		}
+
+		template <typename ParticleType, typename StateArray>
+		AMREX_GPU_DEVICE void operator()(ParticleType &p, StateArray const &state_arr, int i, int j, int k,
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &dx,
+						 amrex::GpuArray<amrex::Real, AMREX_SPACEDIM> const &plo, amrex::Long particle_offset) const
+		{
+			// A simple demonstration of particle creation
+			if (mass_idx + 3 < ParticleType::NReal) {
+				p.pos(0) = plo[0] + (i + 0.5) * dx[0];
+				p.pos(1) = plo[1] + (j + 0.5) * dx[1];
+				p.pos(2) = plo[2] + (k + 0.5) * dx[2];
+
+				// Set particle ID and CPU
+				p.id() = pid_start + particle_offset;
+				p.cpu() = cpu_id;
+
+				// Set particle mass and velocities
+				const amrex::Real cell_volume = AMREX_D_TERM(dx[0], *dx[1], *dx[2]);
+				const amrex::Real cell_density = state_arr(i, j, k, HydroSystem<problem_t>::density_index);
+				const amrex::Real cell_mass = cell_density * cell_volume;
+
+				// Initialize particle properties
+				p.rdata(mass_idx) = 0.5 * cell_mass;
+				p.rdata(mass_idx + 1) = state_arr(i, j, k, HydroSystem<problem_t>::x1Momentum_index) / cell_density;
+				p.rdata(mass_idx + 2) = state_arr(i, j, k, HydroSystem<problem_t>::x2Momentum_index) / cell_density;
+				p.rdata(mass_idx + 3) = state_arr(i, j, k, HydroSystem<problem_t>::x3Momentum_index) / cell_density;
+
+				// Update cell density (remove mass that was given to particle)
+				state_arr(i, j, k, HydroSystem<problem_t>::density_index) = 0.5 * cell_density;
+			}
+		}
+	};
+
+	// Main method to create particles - uses the helper implementation
+	template <typename problem_t, typename ContainerType>
+	static void createParticles(ContainerType *container, int mass_idx, amrex::MultiFab &state, int lev, amrex::Real current_time, amrex::Real dt)
+	{
+		// Use the common implementation with our checker and creator types
+		ParticleCreationImpl::createParticlesImpl<problem_t, ContainerType, ParticleCreationTraits<ParticleType::CIC>::template ParticleChecker,
+							  ParticleCreationTraits<ParticleType::CIC>::template ParticleCreator>(container, mass_idx, state, lev,
+															       current_time, dt);
+	}
+};
+} // namespace quokka
+
 template <> void QuokkaSimulation<BinaryOrbit>::setInitialConditionsOnGrid(quokka::grid const &grid_elem)
 {
 	const amrex::Box &indexRange = grid_elem.indexRange_;
@@ -138,7 +218,7 @@ auto problem_main() -> int
 	double position_error = 0.0;
 	double position_norm = 0.0;
 
-	const int n_particle_expect = 2 + 3 * 3 * 3; // 2 particles from the initial condition, 3*3*3 particles from the creator
+	const int n_particle_expect = 2 + (3 * 3 * 3) * 2; // 2 particles from the initial condition, (3*3*3)*2 particles from the creator
 
 	int status = 0; // Initialize to success
