+++ tests/Gravity3D.in
@@ -28,7 +28,7 @@ stop_time = 5.0
 plotfile_interval = -1
 checkpoint_interval = -1
 
-particles.param1 = 0.007 # create particles at this time, which is step 1
-particles.param2 = 1. # create particles at this time, which is step 3
+particles.param1 = 0.007 # create particles at this time, which is step 2
+particles.param2 = 0.012 # create particles at this time, which is step 3
 
-max_timesteps = 2
+max_timesteps = 3
